refactor!: merge `cloudflare` and `cloudflare-durable`
