+++ src/adapters/cloudflare.ts
@@ -1,57 +1,112 @@
+import type * as CF from "@cloudflare/workers-types";
+import type { DurableObject } from "cloudflare:workers";
 import type { AdapterOptions, AdapterInstance, Adapter } from "../adapter.ts";
+import type * as web from "../../types/web.ts";
 import { toBufferLike } from "../utils.ts";
 import { adapterUtils, getPeers } from "../adapter.ts";
 import { AdapterHookable } from "../hooks.ts";
 import { Message } from "../message.ts";
-import { WSError } from "../error.ts";
 import { Peer, type PeerContext } from "../peer.ts";
+import { StubRequest } from "../_request.ts";
+import { WSError } from "../error.ts";
 
-import type * as _cf from "@cloudflare/workers-types";
+type ResolveDurableStub = (
+  req: CF.Request,
+  env: unknown,
+  context: CF.ExecutionContext,
+) =>
+  | CF.DurableObjectStub
+  | undefined
+  | Promise<CF.DurableObjectStub | undefined>;
 
-// --- types ---
+export interface CloudflareOptions extends AdapterOptions {
+  /**
+   * Durable Object binding name from environment.
+   *
+   * **Note:** This option will be ignored if `resolveDurableStub` is provided.
+   *
+   * @default "$DurableObject"
+   */
+  bindingName?: string;
 
-declare const WebSocketPair: typeof _cf.WebSocketPair;
-declare const Response: typeof _cf.Response;
+  /**
+   * Durable Object instance name.
+   *
+   * **Note:** This option will be ignored if `resolveDurableStub` is provided.
+   *
+   * @default "crossws"
+   */
+  instanceName?: string;
 
-export interface CloudflareAdapter extends AdapterInstance {
-  handleUpgrade(
-    req: _cf.Request,
-    env: unknown,
-    context: _cf.ExecutionContext,
-  ): Promise<_cf.Response>;
+  /**
+   * Custom function that resolves Durable Object binding to handle the WebSocket upgrade.
+   *
+   * **Note:** This option will override `bindingName` and `instanceName`.
+   */
+  resolveDurableStub?: ResolveDurableStub;
 }
 
-export interface CloudflareOptions extends AdapterOptions {}
+// https://developers.cloudflare.com/durable-objects/examples/websocket-hibernation-server/
 
-// --- adapter ---
+const cloudflareAdapter: Adapter<
+  CloudflareDurableAdapter,
+  CloudflareOptions
+> = (opts = {}) => {
+  const hooks = new AdapterHookable(opts);
+  const globalPeers = new Map<
+    string,
+    Set<CloudflareDurablePeer | CloudflareFallbackPeer>
+  >();
 
-// https://developers.cloudflare.com/workers/examples/websockets/
-const cloudflareAdapter: Adapter<CloudflareAdapter, CloudflareOptions> = (
-  options = {},
-) => {
-  const hooks = new AdapterHookable(options);
-  const globalPeers = new Map<string, Set<CloudflarePeer>>();
+  const resolveDurableStub: ResolveDurableStub =
+    opts.resolveDurableStub ||
+    ((_req, env: any, _context): CF.DurableObjectStub | undefined => {
+      const bindingName = opts.bindingName || "$DurableObject";
+      const binding = env[bindingName] as CF.DurableObjectNamespace;
+      if (binding) {
+        const instanceId = binding.idFromName(opts.instanceName || "crossws");
+        return binding.get(instanceId);
+      }
+    });
+
   return {
     ...adapterUtils(globalPeers),
-    handleUpgrade: async (request, env, cfCtx) => {
+    handleUpgrade: async (request, cfEnv, cfCtx) => {
+      // Upgrade request with Durable Object binding
+      const stub = await resolveDurableStub(
+        request as CF.Request,
+        cfEnv,
+        cfCtx,
+      );
+      if (stub) {
+        return stub.fetch(
+          request as CF.Request,
+        ) as unknown as Promise<Response>;
+      }
+
+      // [Fallback] Upgrade request in same Worker
       const { upgradeHeaders, endResponse, context, namespace } =
         await hooks.upgrade(request as unknown as Request);
       if (endResponse) {
-        return endResponse as unknown as _cf.Response;
+        return endResponse as unknown as Response;
       }
-
-      const peers = getPeers(globalPeers, namespace);
-
-      const pair = new WebSocketPair();
+      const peers = getPeers(
+        globalPeers,
+        namespace,
+      ) as Set<CloudflareFallbackPeer>;
+      const pair = new WebSocketPair() as unknown as [
+        CF.WebSocket,
+        CF.WebSocket,
+      ];
       const client = pair[0];
       const server = pair[1];
-      const peer = new CloudflarePeer({
+      const peer = new CloudflareFallbackPeer({
         ws: client,
         peers,
         wsServer: server,
         request: request as unknown as Request,
-        cfEnv: env,
-        cfCtx: cfCtx,
+        cfEnv,
+        cfCtx,
         context,
         namespace,
       });
@@ -76,9 +131,53 @@ const cloudflareAdapter: Adapter<CloudflareAdapter, CloudflareOptions> = (
       // eslint-disable-next-line unicorn/no-null
       return new Response(null, {
         status: 101,
+        webSocket: client as unknown as WebSocket,
+        headers: upgradeHeaders,
+      }) as unknown as Response;
+    },
+    handleDurableInit: async (obj, state, env) => {
+      // placeholder
+    },
+    handleDurableUpgrade: async (obj, request) => {
+      const { upgradeHeaders, endResponse, namespace } = await hooks.upgrade(
+        request as Request,
+      );
+      if (endResponse) {
+        return endResponse;
+      }
+
+      const peers = getPeers(globalPeers, namespace);
+
+      const pair = new WebSocketPair();
+      const client = pair[0];
+      const server = pair[1];
+      const peer = CloudflareDurablePeer._restore(
+        obj,
+        server as unknown as CF.WebSocket,
+        request,
+        namespace,
+      );
+      peers.add(peer);
+      (obj as DurableObjectPub).ctx.acceptWebSocket(server);
+      await hooks.callHook("open", peer);
+
+      // eslint-disable-next-line unicorn/no-null
+      return new Response(null, {
+        status: 101,
         webSocket: client,
         headers: upgradeHeaders,
       });
+    },
+    handleDurableMessage: async (obj, ws, message) => {
+      const peer = CloudflareDurablePeer._restore(obj, ws as CF.WebSocket);
+      await hooks.callHook("message", peer, new Message(message, peer));
+    },
+    handleDurableClose: async (obj, ws, code, reason, wasClean) => {
+      const peer = CloudflareDurablePeer._restore(obj, ws as CF.WebSocket);
+      const peers = getPeers(globalPeers, peer.namespace);
+      peers.delete(peer);
+      const details = { code, reason, wasClean };
+      await hooks.callHook("close", peer, details);
     },
   };
 };
@@ -87,13 +186,97 @@ export default cloudflareAdapter;
 
 // --- peer ---
 
-class CloudflarePeer extends Peer<{
-  ws: _cf.WebSocket;
+class CloudflareDurablePeer extends Peer<{
+  ws: AugmentedWebSocket;
   request: Request;
-  peers: Set<CloudflarePeer>;
-  wsServer: _cf.WebSocket;
+  peers?: never;
+  durable: DurableObjectPub;
+  namespace: string;
+}> {
+  override get peers() {
+    return new Set(
+      this.#getwebsockets().map((ws) =>
+        CloudflareDurablePeer._restore(this._internal.durable, ws),
+      ),
+    );
+  }
+
+  #getwebsockets() {
+    return this._internal.durable.ctx.getWebSockets() as unknown as (typeof this._internal.ws)[];
+  }
+
+  send(data: unknown) {
+    return this._internal.ws.send(toBufferLike(data));
+  }
+
+  override subscribe(topic: string): void {
+    super.subscribe(topic);
+    const state = getAttachedState(this._internal.ws);
+    if (!state.t) {
+      state.t = new Set();
+    }
+    state.t.add(topic);
+    setAttachedState(this._internal.ws, state);
+  }
+
+  publish(topic: string, data: unknown): void {
+    const websockets = this.#getwebsockets();
+    if (websockets.length < 2 /* 1 is self! */) {
+      return;
+    }
+    const dataBuff = toBufferLike(data);
+    for (const ws of websockets) {
+      if (ws === this._internal.ws) {
+        continue;
+      }
+      const state = getAttachedState(ws);
+      if (state.t?.has(topic)) {
+        ws.send(dataBuff);
+      }
+    }
+  }
+
+  close(code?: number, reason?: string) {
+    this._internal.ws.close(code, reason);
+  }
+
+  static _restore(
+    durable: DurableObject,
+    ws: AugmentedWebSocket,
+    request?: Request | CF.Request,
+    namespace?: string,
+  ): CloudflareDurablePeer {
+    let peer = ws._crosswsPeer;
+    if (peer) {
+      return peer;
+    }
+    const state = (ws.deserializeAttachment() || {}) as AttachedState;
+    peer = ws._crosswsPeer = new CloudflareDurablePeer({
+      ws: ws as CF.WebSocket,
+      request:
+        (request as Request | undefined) || new StubRequest(state.u || ""),
+      namespace: namespace || state.n || "" /* later throws error if empty */,
+      durable: durable as DurableObjectPub,
+    });
+    if (state.i) {
+      peer._id = state.i;
+    }
+    if (request?.url) {
+      state.u = request.url;
+    }
+    state.i = peer.id;
+    setAttachedState(ws, state);
+    return peer;
+  }
+}
+
+class CloudflareFallbackPeer extends Peer<{
+  ws: CF.WebSocket;
+  request: Request;
+  peers: Set<CloudflareFallbackPeer>;
+  wsServer: CF.WebSocket;
   cfEnv: unknown;
-  cfCtx: _cf.ExecutionContext;
+  cfCtx: CF.ExecutionContext;
   context: PeerContext;
   namespace: string;
 }> {
@@ -109,4 +292,78 @@ class CloudflarePeer extends Peer<{
   close(code?: number, reason?: string) {
     this._internal.ws.close(code, reason);
   }
+}
+
+// -- attached state utils ---
+
+function getAttachedState(ws: AugmentedWebSocket): AttachedState {
+  let state = ws._crosswsState;
+  if (state) {
+    return state;
+  }
+  state = (ws.deserializeAttachment() as AttachedState) || {};
+  ws._crosswsState = state;
+  return state;
+}
+
+function setAttachedState(ws: AugmentedWebSocket, state: AttachedState) {
+  ws._crosswsState = state;
+  ws.serializeAttachment(state);
+}
+
+// --- types ---
+
+declare class DurableObjectPub extends DurableObject {
+  public ctx: DurableObject["ctx"];
+  public env: unknown;
+}
+
+type AugmentedWebSocket = CF.WebSocket & {
+  _crosswsPeer?: CloudflareDurablePeer;
+  _crosswsState?: AttachedState;
+};
+
+/** Max serialized limit: 2048 bytes (512..2048 characters) */
+type AttachedState = {
+  /** Subscribed topics */
+  t?: Set<string>;
+  /** Peer id */
+  i?: string;
+  /** Request url */
+  u?: string;
+  /** Connection namespace */
+  n?: string;
+};
+
+export interface CloudflareDurableAdapter extends AdapterInstance {
+  handleUpgrade(
+    req: Request | CF.Request,
+    env: unknown,
+    context: CF.ExecutionContext,
+  ): Promise<Response>;
+
+  handleDurableInit(
+    obj: DurableObject,
+    state: DurableObjectState,
+    env: unknown,
+  ): void;
+
+  handleDurableUpgrade(
+    obj: DurableObject,
+    req: Request | CF.Request,
+  ): Promise<Response>;
+
+  handleDurableMessage(
+    obj: DurableObject,
+    ws: WebSocket | CF.WebSocket | web.WebSocket,
+    message: ArrayBuffer | string,
+  ): Promise<void>;
+
+  handleDurableClose(
+    obj: DurableObject,
+    ws: WebSocket | CF.WebSocket | web.WebSocket,
+    code: number,
+    reason: string,
+    wasClean: boolean,
+  ): Promise<void>;
 }
